requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
SQLAlchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
flask==3.0.2
flask-sqlalchemy==3.1.1 